Move test since it doesn't require a running TypeDB server
