+++ kglib/utils/graph/thing/queries_to_networkx_graph_it.py
@@ -28,133 +28,8 @@ from typedb.client import *
 
 from kglib.utils.graph.test.case import GraphTestCase
 from kglib.utils.graph.thing.queries_to_networkx_graph import build_graph_from_queries
-from kglib.utils.typedb.object.thing import Thing, build_thing
+from kglib.utils.typedb.object.thing import build_thing
 from kglib.utils.typedb.test.base import TypeDBServer
-from kglib.utils.typedb.test.mock.answer import MockConceptMap
-from kglib.utils.typedb.test.mock.concept import MockType, MockAttributeType, MockThing, MockAttribute
-
-TEST_URI = "localhost:1729"
-
-
-def mock_sampler(input_iter):
-    return input_iter
-
-
-class ITBuildGraphFromQueries(GraphTestCase):
-    def test_graph_is_built_as_expected(self):
-        g1 = nx.MultiDiGraph()
-        g1.add_node('x')
-
-        g2 = nx.MultiDiGraph()
-        g2.add_node('x')
-        g2.add_node('n')
-        g2.add_edge('x', 'n', type='has')
-
-        g3 = nx.MultiDiGraph()
-        g3.add_node('x')
-        g3.add_node('r')
-        g3.add_node('y')
-        g3.add_edge('r', 'x', type='child')
-        g3.add_edge('r', 'y', type='parent')
-
-        query_sampler_variable_graph_tuples = [('match $x iid V123;', mock_sampler, g1),
-                                               ('match $x iid V123, has name $n;', mock_sampler, g2),
-                                               ('match $x iid V123; $r(child: $x, parent: $y);', mock_sampler, g3),
-                                               # TODO Add functionality for loading schema at a later date
-                                               # ('match $x sub person; $x sub $type;', g4),
-                                               # ('match $x sub $y;', g5),
-                                               ]
-
-        class MockTransaction:
-
-            def query(self):
-                return MockQueryManager()
-
-        class MockQueryManager:
-            def match(self, query):
-
-                if query == 'match $x iid V123;':
-                    return [MockConceptMap({'x': MockThing('V123', MockType('V4123', 'person', 'ENTITY'))})]
-                elif query == 'match $x iid V123, has name $n;':
-                    return [
-                        MockConceptMap({
-                            'x': MockThing('V123', MockType('V4123', 'person', 'ENTITY')),
-                            'n': MockAttribute('V987', 'Bob', MockAttributeType('V555', 'name', 'ATTRIBUTE',
-                                                                                AttributeType.ValueType.STRING))
-                        })]
-                elif query == 'match $x iid V123; $r(child: $x, parent: $y);':
-                    return [
-                        MockConceptMap({
-                            'x': MockThing('V123', MockType('V4123', 'person', 'ENTITY')),
-                            'y': MockThing('V123', MockType('V4123', 'person', 'ENTITY')),
-                            'r': MockThing('V567', MockType('V9876', 'parentship', 'RELATION'))
-                        })]
-                else:
-                    raise NotImplementedError
-
-        mock_tx = MockTransaction()
-
-        combined_graph = build_graph_from_queries(query_sampler_variable_graph_tuples, mock_tx)
-
-        person_exp = Thing('V123', 'person', 'entity')
-        parentship_exp = Thing('V567', 'parentship', 'relation')
-        name_exp = Thing('V987', 'name', 'attribute', value_type=AttributeType.ValueType.STRING, value='Bob')
-        expected_combined_graph = nx.MultiDiGraph()
-        expected_combined_graph.add_node(person_exp, type='person')
-        expected_combined_graph.add_node(name_exp, type='name', value_type=AttributeType.ValueType.STRING, value='Bob')
-        expected_combined_graph.add_node(parentship_exp, type='parentship')
-        expected_combined_graph.add_edge(parentship_exp, person_exp, type='child')
-        expected_combined_graph.add_edge(parentship_exp, person_exp, type='parent')
-        expected_combined_graph.add_edge(person_exp, name_exp, type='has')
-
-        self.assertGraphsEqual(expected_combined_graph, combined_graph)
-
-    def test_warning_given_when_one_query_gives_no_results(self):
-        g1 = nx.MultiDiGraph()
-        g1.add_node('x')
-        g2 = nx.MultiDiGraph()
-        g2.add_node('y')
-        query_sampler_variable_graph_tuples = [('match $x iid V123;', mock_sampler, g1),
-                                               ('match $y iid V123;', mock_sampler, g2)]
-
-        class MockTransaction:
-            def query(self):
-                return MockQueryManager()
-
-        class MockQueryManager:
-            def match(self, query):
-                if query == 'match $x iid V123;':
-                    return [MockConceptMap({'x': MockThing('V123', MockType('V4123', 'person', 'ENTITY'))})]
-                elif query == 'match $y iid V123;':
-                    return []
-
-        mock_tx = MockTransaction()
-
-        with self.assertWarns(UserWarning) as context:
-            build_graph_from_queries(query_sampler_variable_graph_tuples, mock_tx)
-
-        self.assertEqual(f'There were no results for query: \n\"match $y iid V123;\"\nand so nothing will be added to the graph for this query', str(context.warning))
-
-    def test_exception_is_raised_when_there_are_no_results_for_any_query(self):
-        g1 = nx.MultiDiGraph()
-        g1.add_node('x')
-        query_sampler_variable_graph_tuples = [('match $x iid V123;', mock_sampler, g1)]
-
-        class QueryManagerEmpty:
-            def match(self, query):
-                return []
-
-        class MockTransactionEmpty:
-            def query(self):
-                return QueryManagerEmpty()
-
-        mock_tx = MockTransactionEmpty()
-
-        with self.assertRaises(RuntimeError) as context:
-            build_graph_from_queries(query_sampler_variable_graph_tuples, mock_tx)
-
-        self.assertEqual(f'The graph from queries: {[query_sampler_variable_graph_tuple[0] for query_sampler_variable_graph_tuple in query_sampler_variable_graph_tuples]}\n'
-                         f'could not be created, since none of these queries returned results', str(context.exception))
 
 
 class ITBuildGraphFromQueriesWithRealTypeDB(GraphTestCase):
